Update favicon, title, Organise folders
