+++ README.md
@@ -1,4 +1,4 @@
-# Inverse Hangman
+# Happy Tree Man: Reverse Hangman
 
 Save the hangman before all his limbs get forcefully removed by cute animals. Play it on your browser. Designed and developed as part of Tiktok Youth Camp 2022.
