+++ src/components/App.js
@@ -1,5 +1,5 @@
-import logo from './logo.svg';
-import './App.css';
+import logo from '../assets/logo.svg';
+import '../assets/App.css';
 
 function App() {
   return (
